clean up and abstraction
